updating the model_builder.py
removing functions that are not used to build a model
